- Made previews as a PreviewRepo, delivered via DI
