+++ app/src/main/java/space/taran/arknavigator/mvp/presenter/adapter/FoldersWalker.kt
@@ -1,18 +1,23 @@
 package space.taran.arknavigator.mvp.presenter.adapter
 
+import space.taran.arknavigator.mvp.model.repo.PreviewRepo
 import space.taran.arknavigator.ui.fragments.utils.Preview
 import space.taran.arknavigator.mvp.view.item.FileItemView
 import space.taran.arknavigator.utils.findLongestCommonPrefix
 import java.nio.file.Path
+import javax.inject.Inject
 
 class FoldersWalker(paths: List<Path>, onClick: ItemClickHandler<Path>)
     : ItemsReversiblePresenter<Path, Path, FileItemView>(
         findLongestCommonPrefix(paths), paths, onClick) {
 
+    @Inject
+    lateinit var previewRepo: PreviewRepo
+
     override fun bindView(view: FileItemView) {
         val path = items()[view.position()]
         view.setText(path.fileName.toString())
 
-        view.setIcon(Preview.provide(path))
+        view.setIcon(previewRepo.providePreview(path))
     }
 }
