+++ app/src/main/java/space/taran/arknavigator/ui/adapter/FolderPicker.kt
@@ -3,6 +3,7 @@ package space.taran.arknavigator.ui.adapter
 import space.taran.arknavigator.databinding.DialogRootsNewBinding
 import space.taran.arknavigator.mvp.presenter.adapter.FoldersWalker
 import space.taran.arknavigator.mvp.presenter.adapter.ItemClickHandler
+import space.taran.arknavigator.ui.App
 import java.nio.file.Path
 import kotlin.io.path.ExperimentalPathApi
 import kotlin.io.path.listDirectoryEntries
@@ -11,8 +12,10 @@ import kotlin.io.path.listDirectoryEntries
 class FolderPicker(
     paths: List<Path>,
     handler: ItemClickHandler<Path>,
-    private val dialogBinding: DialogRootsNewBinding)
-    : FilesReversibleRVAdapter<Path, Path>(FoldersWalker(paths, handler)) {
+    private val dialogBinding: DialogRootsNewBinding
+) : FilesReversibleRVAdapter<Path, Path>(FoldersWalker(paths, handler).apply {
+    App.instance.appComponent.inject(this)
+}) {
 
     init {
         dialogBinding.rvRootsDialog.adapter = this
